Move velocity formulas to calculations file
